+++ src/data.txt
@@ -1,3 +1,2 @@
-T|true|adsfas
 T|true|asdfa
 D|false|asdfas |2022-12-11
