+++ src/main/java/duke/commands/AddDeadline.java
@@ -1,16 +1,22 @@
-public class AddEvents extends Command{
+package duke.commands;
+
+import duke.tasklist.DukeList;
+import duke.task.Deadlines;
+import duke.ui.Ui;
+
+public class AddDeadline extends Command{
 
     private String msg;
     private String date;
 
-    public AddEvents(String msg, String date){
+    public AddDeadline(String msg, String date){
         this.msg = msg;
         this.date = date;
     }
 
     @Override
     public void execute(Ui ui, DukeList list) {
-        Events ans = new Events(msg, date);
+        Deadlines ans = new Deadlines(msg, date);
         list.add(ans);
     }
