+++ src/main/java/duke/commands/AddEvents.java
@@ -1,16 +1,22 @@
-public class AddDeadline extends Command{
+package duke.commands;
+
+import duke.tasklist.DukeList;
+import duke.task.Events;
+import duke.ui.Ui;
+
+public class AddEvents extends Command{
 
     private String msg;
     private String date;
 
-    public AddDeadline(String msg, String date){
+    public AddEvents(String msg, String date){
         this.msg = msg;
         this.date = date;
     }
 
     @Override
     public void execute(Ui ui, DukeList list) {
-        Deadlines ans = new Deadlines(msg, date);
+        Events ans = new Events(msg, date);
         list.add(ans);
     }
