+++ src/main/java/duke/commands/AddToDos.java
@@ -1,3 +1,9 @@
+package duke.commands;
+
+import duke.tasklist.DukeList;
+import duke.task.ToDos;
+import duke.ui.Ui;
+
 public class AddToDos extends Command{
 
     private String body;
