+++ src/main/java/duke/commands/Command.java
@@ -1,3 +1,8 @@
+package duke.commands;
+
+import duke.tasklist.DukeList;
+import duke.ui.Ui;
+
 public abstract class Command {
 
     public abstract void execute(Ui ui, DukeList list);
