+++ src/main/java/duke/commands/ExitCommand.java
@@ -1,3 +1,8 @@
+package duke.commands;
+
+import duke.tasklist.DukeList;
+import duke.ui.Ui;
+
 public class ExitCommand extends Command{
