+++ src/main/java/duke/commands/TodayTask.java
@@ -1,7 +1,13 @@
-public class TryAgain extends Command{
+package duke.commands;
+
+import duke.tasklist.DukeList;
+import duke.ui.Ui;
+
+public class TodayTask extends Command{
 
     @Override
     public void execute(Ui ui, DukeList list) {
+        list.todayTask();
     }
 
     @Override
