+++ src/main/java/duke/commands/TryAgain.java
@@ -1,8 +1,12 @@
-public class TodayTask extends Command{
+package duke.commands;
+
+import duke.tasklist.DukeList;
+import duke.ui.Ui;
+
+public class TryAgain extends Command{
 
     @Override
     public void execute(Ui ui, DukeList list) {
-        list.todayTask();
     }
 
     @Override
