+++ src/main/java/duke/commands/UnMarkCommand.java
@@ -1,14 +1,19 @@
-public class DeleteCommand extends Command{
+package duke.commands;
+
+import duke.tasklist.DukeList;
+import duke.ui.Ui;
+
+public class UnMarkCommand extends Command{
 
     private int index;
 
-    public DeleteCommand(int x){
+    public UnMarkCommand(int x){
         index = x;
     }
 
     @Override
     public void execute(Ui ui, DukeList tasks){
-        tasks.delete(index);
+        tasks.unmark(index);
     }
 
     @Override
