+++ src/main/java/duke/exceptions/InvalidDateException.java
@@ -1,3 +1,5 @@
+package duke.exceptions;
+
 public class InvalidDateException extends DukeException{
 
     public InvalidDateException(){
