+++ src/main/java/duke/exceptions/MissingArgumentException.java
@@ -1,3 +1,5 @@
+package duke.exceptions;
+
 public class MissingArgumentException extends DukeException{
 
     public MissingArgumentException(){
