+++ src/main/java/duke/exceptions/WrongFormatException.java
@@ -1,3 +1,5 @@
+package duke.exceptions;
+
 public class WrongFormatException extends DukeException{
 
     public WrongFormatException(){
