+++ src/main/java/duke/task/Deadlines.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 import java.time.LocalDate;
 
 public class Deadlines extends Task{
