+++ src/main/java/duke/task/Events.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 import java.time.LocalDate;
 
 public class Events extends Task {
