+++ src/main/java/duke/test.java
@@ -1,4 +1,5 @@
-import java.io.File;
+package duke;
+
 import java.io.IOException;
 import java.util.Scanner;
