+++ text-ui-test/EXPECTED.TXT
@@ -1,59 +1,59 @@
-Duke: Hello! I'm Duke
+duke: Hello! I'm duke
       What can I do for you?
 
-Duke: Got it. I've added this task:
+duke: Got it. I've added this task:
       [T][ ]  read book
       Now you have 1 tasks in the list.
 
 
-Duke: Got it. I've added this task:
+duke: Got it. I've added this task:
       [D][ ]  return book  (by:  Sunday)
       Now you have 2 tasks in the list.
 
 
-Duke: Got it. I've added this task:
+duke: Got it. I've added this task:
       [E][ ]  project meeting  (at:  Mon 2-4pm)
       Now you have 3 tasks in the list.
 
-Duke: Nice! I've marked this task as done:
+duke: Nice! I've marked this task as done:
       [T][X]  read book
-Duke: Nice! I've marked this task as done:
+duke: Nice! I've marked this task as done:
       [D][X]  return book  (by:  Sunday)
-Duke: Nice! I've marked this task as done:
+duke: Nice! I've marked this task as done:
       [E][X]  project meeting  (at:  Mon 2-4pm)
 
-Duke:
+duke:
 Here are the tasks in your list:
 1. [T][X]  read book
 2. [D][X]  return book  (by:  Sunday)
 3. [E][X]  project meeting  (at:  Mon 2-4pm)
 
-Duke: OK, I've marked this task as not done yet:
+duke: OK, I've marked this task as not done yet:
       [T][ ]  read book
 
-Duke:
+duke:
 Here are the tasks in your list:
 1. [T][ ]  read book
 2. [D][X]  return book  (by:  Sunday)
 3. [E][X]  project meeting  (at:  Mon 2-4pm)
 
-Duke: OK, I've marked this task as not done yet:
+duke: OK, I've marked this task as not done yet:
       [D][ ]  return book  (by:  Sunday)
 
-Duke:
+duke:
 Here are the tasks in your list:
 1. [T][ ]  read book
 2. [D][ ]  return book  (by:  Sunday)
 3. [E][X]  project meeting  (at:  Mon 2-4pm)
 
-Duke: OK, I've marked this task as not done yet:
+duke: OK, I've marked this task as not done yet:
       [E][ ]  project meeting  (at:  Mon 2-4pm)
 
-Duke:
+duke:
 Here are the tasks in your list:
 1. [T][ ]  read book
 2. [D][ ]  return book  (by:  Sunday)
 3. [E][ ]  project meeting  (at:  Mon 2-4pm)
 
 
-Duke:  Bye. Hope to see you again soon!
+duke:  Bye. Hope to see you again soon!
